Implement Level 1. Greet, Echo, Exit.
