+++ src/test/java/seedu/duke/DukeTest.java
@@ -1,0 +1,15 @@
+package seedu.duke;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+/**
+ * Tests Duke.
+ */
+public class DukeTest {
+    @Test
+    public void dummyTest() {
+        assertEquals(2, 2);
+    }
+}
